ava-backend/requirements.txt
Flask
python-dotenv
openai>=1.0.0
flask_cors
